requirements.txt
colorgram.py==1.2.0
pillow==11.1.0
